+++ src/model/person/person.cpp
@@ -1,9 +1,9 @@
 
 #include "person.h"
 
-const int Person::DEFAULT_TAX_ID = 999999999;
+const unsigned long Person::DEFAULT_TAX_ID = 999999999;
 
-Person::Person(std::string name, int taxID, Credential credential, PersonRole role) :
+Person::Person(std::string name, unsigned long taxID, Credential credential, PersonRole role) :
         _name(std::move(name)), _taxID{taxID}, _credential{std::move(credential) }, _logged(false),
         _role(role) {
     if (credential.isReserved()) throw InvalidCredential();
@@ -13,7 +13,7 @@ std::string Person::getName() const {
     return _name;
 }
 
-int Person::getTaxId() const {
+unsigned long Person::getTaxId() const {
     return _taxID;
 }
 
@@ -48,7 +48,7 @@ void Person::setLogged(bool logged) {
     _logged = logged;
 }
 
-void Person::setTaxID(int taxID) {
+void Person::setTaxID(unsigned long taxID) {
     _taxID = taxID;
 }
