+++ src/model/person/person.h
@@ -11,7 +11,7 @@
 #include <iostream>
 
 /**
- * The enum with the possible person role.
+ * The enum with the possible person roles.
  */
 enum class PersonRole {
     WORKER,
@@ -66,7 +66,7 @@ public:
      * @param taxID the taxpayer identification number
      * @param credential the login credentials (username and password)
      */
-    Person(std::string name, int taxID, Credential credential, PersonRole role);
+    Person(std::string name, unsigned long taxID, Credential credential, PersonRole role);
 
     /**
      * Destructs the person object.
@@ -85,7 +85,7 @@ public:
      *
      * @return the taxpayer identification number
      */
-    int getTaxId() const;
+    unsigned long getTaxId() const;
 
     /**
      * Gets the person login credentials.
@@ -120,7 +120,7 @@ public:
      *
      * @param taxID the taxpayer identification number
      */
-    void setTaxID(int taxID);
+    void setTaxID(unsigned long taxID);
 
     /**
      * Sets the person login credentials.
@@ -165,7 +165,7 @@ public:
     /**
      * The default taxpayer identification number.
      */
-    static const int DEFAULT_TAX_ID;
+    static const unsigned long DEFAULT_TAX_ID;
 
 private:
     /**
@@ -176,7 +176,7 @@ private:
     /**
      * The person taxpayer identification number.
      */
-    int _taxID;
+    unsigned long _taxID;
 
     /**
      * The person login credentials.
@@ -200,11 +200,11 @@ protected:
  */
 struct PersonSmaller{
     /**
-     * Checks if a person is smaller than other.
+     * Checks if a person should be shown before another.
      *
      * @param p1 the first person to compare with
      * @param p2 the second person to compare with
-     * @return true, if p1 is less than p2, according to alphabetical order; false, otherwise
+     * @return true, if p1 is less than p2, according to defined equality operator
      */
     bool operator()(const Person* p1, const Person* p2) const{
         return *p1 < *p2;
